requirements.txt
pandas==1.3.3
python_docx==0.8.11
streamlit==1.38.0
~treamlit==1.29.0
